fix no storage

+++ test/conftest.py
@@ -3,6 +3,7 @@ import tempfile
 
 import pytest
 
+# whether to remove storage files after a test has finished
 REMOVE_STORE = True
 
 
@@ -10,7 +11,8 @@ REMOVE_STORE = True
 def hdf5_file():
     """Generate a temporary hdf5 file."""
     store_file = tempfile.mkstemp(suffix=".hdf5")[1]
-    yield store_file
-    if REMOVE_STORE:
-        if os.path.exists(store_file):
+    try:
+        yield store_file
+    finally:
+        if REMOVE_STORE and os.path.exists(store_file):
             os.remove(store_file)
